Restore session message rendering helpers
