+++ vtcode-core/src/ui/tui/session.rs
@@ -1,4 +1,4 @@
-use std::{cmp::min, fmt::Write, mem, time::Instant};
+use std::{cmp::min, mem, time::Instant};
 
 use ansi_to_tui::IntoText;
 use anstyle::{AnsiColor, Color as AnsiColorEnum, RgbColor};
@@ -21,18 +21,21 @@ use unicode_segmentation::UnicodeSegmentation;
 use unicode_width::UnicodeWidthStr;
 
 use super::types::{
-    InlineCommand, InlineEvent, InlineHeaderContext, InlineHeaderHighlight, InlineListItem,
-    InlineListSearchConfig, InlineListSelection, InlineMessageKind, InlineSegment, InlineTextStyle,
-    InlineTheme, SecurePromptConfig,
+    InlineCommand, InlineEvent, InlineHeaderContext, InlineListItem, InlineListSearchConfig,
+    InlineListSelection, InlineMessageKind, InlineSegment, InlineTextStyle, InlineTheme,
+    SecurePromptConfig,
 };
 use crate::config::constants::{prompts, ui};
 
 mod file_palette;
+mod header;
 mod input;
 mod message;
 mod modal;
+mod navigation;
 mod prompt_palette;
 mod queue;
+mod slash;
 mod slash_palette;
 mod transcript;
 
@@ -45,12 +48,14 @@ use self::modal::{
 };
 use self::prompt_palette::{PromptPalette, extract_prompt_reference};
 use self::queue::QueueOverlay;
-use self::slash_palette::{SlashPalette, SlashPaletteUpdate, command_prefix, command_range};
+use self::slash_palette::SlashPalette;
 use self::transcript::{CachedMessage, TranscriptReflowCache};
+#[cfg(test)]
+use super::types::InlineHeaderHighlight;
 use crate::prompts::CustomPromptRegistry;
 #[cfg(test)]
 use crate::tools::PlanSummary;
-use crate::tools::{PlanCompletionState, PlanStep, StepStatus, TaskPlan};
+use crate::tools::TaskPlan;
 
 const USER_PREFIX: &str = "";
 const PLACEHOLDER_COLOR: RgbColor = RgbColor(0x88, 0x88, 0x88);
@@ -510,780 +515,9 @@ impl Session {
         self.render_prompt_palette(frame, viewport);
     }
 
-    fn render_header(&self, frame: &mut Frame<'_>, area: Rect, lines: &[Line<'static>]) {
-        frame.render_widget(Clear, area);
-        if area.height == 0 || area.width == 0 {
-            return;
-        }
-
-        let paragraph = self.build_header_paragraph(lines);
-
-        frame.render_widget(paragraph, area);
-    }
-
-    fn header_lines(&self) -> Vec<Line<'static>> {
-        let mut lines = vec![self.header_title_line(), self.header_meta_line()];
-
-        // Prioritize suggestions when input is empty or starts with /
-        if self.should_show_suggestions() {
-            if let Some(suggestions) = self.header_suggestions_line() {
-                lines.push(suggestions);
-            }
-        } else if let Some(highlights) = self.header_highlights_line() {
-            lines.push(highlights);
-        }
-
-        lines.truncate(3);
-        lines
-    }
-
-    fn header_height_from_lines(&self, width: u16, lines: &[Line<'static>]) -> u16 {
-        if width == 0 {
-            return self.header_rows.max(ui::INLINE_HEADER_HEIGHT);
-        }
-
-        let paragraph = self.build_header_paragraph(lines);
-        let measured = paragraph.line_count(width);
-        let resolved = u16::try_from(measured).unwrap_or(u16::MAX);
-        // Limit to max 3 lines to accommodate suggestions
-        resolved.min(3).max(ui::INLINE_HEADER_HEIGHT)
-    }
-
-    fn build_header_paragraph(&self, lines: &[Line<'static>]) -> Paragraph<'static> {
-        let block = Block::default()
-            .title(self.header_block_title())
-            .borders(Borders::ALL)
-            .border_type(BorderType::Rounded)
-            .style(self.default_style());
-
-        Paragraph::new(lines.to_vec())
-            .style(self.default_style())
-            .wrap(Wrap { trim: true })
-            .block(block)
-    }
-
-    #[cfg(test)]
-    fn header_height_for_width(&self, width: u16) -> u16 {
-        let lines = self.header_lines();
-        self.header_height_from_lines(width, &lines)
-    }
-
-    fn render_navigation(&mut self, frame: &mut Frame<'_>, area: Rect) {
-        frame.render_widget(Clear, area);
-        if area.height == 0 || area.width == 0 {
-            return;
-        }
-
-        let block = Block::default()
-            .title(self.navigation_block_title())
-            .borders(Borders::LEFT)
-            .border_type(BorderType::Plain)
-            .style(self.default_style())
-            .border_style(self.border_style());
-        let inner = block.inner(area);
-        if inner.height == 0 {
-            frame.render_widget(block, area);
-            return;
-        }
-
-        let items = self.navigation_items();
-        let item_count = items.len();
-        let viewport = inner.height as usize;
-
-        if self.should_show_plan() {
-            if item_count == 0 {
-                self.navigation_state.select(None);
-                *self.navigation_state.offset_mut() = 0;
-            } else if let Some(selected) = self.plan_selected_index() {
-                self.navigation_state.select(Some(selected));
-                let max_offset = item_count.saturating_sub(viewport);
-                let desired_offset = selected.saturating_sub(viewport.saturating_sub(1));
-                *self.navigation_state.offset_mut() = desired_offset.min(max_offset);
-            } else {
-                self.navigation_state.select(None);
-                *self.navigation_state.offset_mut() = 0;
-            }
-        } else if self.lines.is_empty() {
-            self.navigation_state.select(None);
-            *self.navigation_state.offset_mut() = 0;
-        } else {
-            let last_index = self.lines.len().saturating_sub(1);
-            self.navigation_state.select(Some(last_index));
-            let max_offset = item_count.saturating_sub(viewport);
-            *self.navigation_state.offset_mut() = max_offset;
-        }
-
-        let list = List::new(items)
-            .block(block)
-            .style(self.default_style())
-            .highlight_style(self.navigation_highlight_style());
-
-        frame.render_stateful_widget(list, area, &mut self.navigation_state);
-    }
-
-    fn header_block_title(&self) -> Line<'static> {
-        let fallback = InlineHeaderContext::default();
-        let version = if self.header_context.version.trim().is_empty() {
-            fallback.version
-        } else {
-            self.header_context.version.clone()
-        };
-
-        let prompt = format!(
-            "{}{} ",
-            ui::HEADER_VERSION_PROMPT,
-            ui::HEADER_VERSION_PREFIX
-        );
-        let version_text = format!(
-            "{}{}{}",
-            ui::HEADER_VERSION_LEFT_DELIMITER,
-            version.trim(),
-            ui::HEADER_VERSION_RIGHT_DELIMITER
-        );
-
-        let prompt_style = self.section_title_style();
-        let version_style = self.header_secondary_style().add_modifier(Modifier::DIM);
-
-        Line::from(vec![
-            Span::styled(prompt, prompt_style),
-            Span::styled(version_text, version_style),
-        ])
-    }
-
-    fn header_title_line(&self) -> Line<'static> {
-        // First line: badge-style provider + model + reasoning summary
-        let mut spans = Vec::new();
-
-        let provider = self.header_provider_short_value();
-        let model = self.header_model_short_value();
-        let reasoning = self.header_reasoning_short_value();
-
-        if !provider.is_empty() {
-            let badge = format!("[{}]", provider.to_uppercase());
-            let mut style = self.header_primary_style();
-            style = style.add_modifier(Modifier::BOLD);
-            spans.push(Span::styled(badge, style));
-        }
-
-        if !model.is_empty() {
-            if !spans.is_empty() {
-                spans.push(Span::raw(" "));
-            }
-            let mut style = self.header_primary_style();
-            style = style.add_modifier(Modifier::ITALIC);
-            spans.push(Span::styled(model, style));
-        }
-
-        if !reasoning.is_empty() {
-            if !spans.is_empty() {
-                spans.push(Span::raw(" "));
-            }
-            let mut style = self.header_secondary_style();
-            style = style.add_modifier(Modifier::ITALIC | Modifier::DIM);
-            spans.push(Span::styled(format!("({})", reasoning), style));
-        }
-
-        if spans.is_empty() {
-            spans.push(Span::raw(String::new()));
-        }
-
-        Line::from(spans)
-    }
-
-    fn header_provider_value(&self) -> String {
-        let trimmed = self.header_context.provider.trim();
-        if trimmed.is_empty() {
-            InlineHeaderContext::default().provider
-        } else {
-            self.header_context.provider.clone()
-        }
-    }
-
-    fn header_model_value(&self) -> String {
-        let trimmed = self.header_context.model.trim();
-        if trimmed.is_empty() {
-            InlineHeaderContext::default().model
-        } else {
-            self.header_context.model.clone()
-        }
-    }
-
-    fn header_mode_label(&self) -> String {
-        let trimmed = self.header_context.mode.trim();
-        if trimmed.is_empty() {
-            InlineHeaderContext::default().mode
-        } else {
-            self.header_context.mode.clone()
-        }
-    }
-
-    fn header_mode_short_label(&self) -> String {
-        let full = self.header_mode_label();
-        let value = full.trim();
-        if value.eq_ignore_ascii_case(ui::HEADER_MODE_AUTO) {
-            return "Auto".to_string();
-        }
-        if value.eq_ignore_ascii_case(ui::HEADER_MODE_INLINE) {
-            return "Inline".to_string();
-        }
-        if value.eq_ignore_ascii_case(ui::HEADER_MODE_ALTERNATE) {
-            return "Alternate".to_string();
-        }
-        let compact = value
-            .strip_suffix(ui::HEADER_MODE_FULL_AUTO_SUFFIX)
-            .unwrap_or(value)
-            .trim();
-        compact.to_string()
-    }
-
-    fn header_reasoning_value(&self) -> Option<String> {
-        let trimmed = self.header_context.reasoning.trim();
-        let value = if trimmed.is_empty() {
-            InlineHeaderContext::default().reasoning
-        } else {
-            self.header_context.reasoning.clone()
-        };
-        if value.trim().is_empty() {
-            None
-        } else {
-            Some(value)
-        }
-    }
-
-    fn strip_prefix<'a>(value: &'a str, prefix: &str) -> &'a str {
-        value.strip_prefix(prefix).unwrap_or(value)
-    }
-
-    fn header_provider_short_value(&self) -> String {
-        let value = self.header_provider_value();
-        Self::strip_prefix(&value, ui::HEADER_PROVIDER_PREFIX)
-            .trim()
-            .to_string()
-    }
-
-    fn header_model_short_value(&self) -> String {
-        let value = self.header_model_value();
-        Self::strip_prefix(&value, ui::HEADER_MODEL_PREFIX)
-            .trim()
-            .to_string()
-    }
-
-    fn header_reasoning_short_value(&self) -> String {
-        let value = self.header_reasoning_value().unwrap_or_else(String::new);
-        Self::strip_prefix(&value, ui::HEADER_REASONING_PREFIX)
-            .trim()
-            .to_string()
-    }
-
-    fn header_chain_values(&self) -> Vec<String> {
-        let defaults = InlineHeaderContext::default();
-        let fields = [
-            (
-                &self.header_context.workspace_trust,
-                defaults.workspace_trust.clone(),
-            ),
-            (&self.header_context.tools, defaults.tools.clone()),
-            (&self.header_context.git, defaults.git.clone()),
-            // Removed MCP info from header as requested
-        ];
-
-        fields
-            .into_iter()
-            .filter_map(|(value, fallback)| {
-                let mut selected = if value.trim().is_empty() {
-                    fallback
-                } else {
-                    value.clone()
-                };
-                let trimmed = selected.trim();
-                if trimmed.is_empty() {
-                    return None;
-                }
-
-                if let Some(body) = trimmed.strip_prefix(ui::HEADER_TRUST_PREFIX) {
-                    selected = format!("Trust {}", body.trim());
-                    return Some(selected);
-                }
-
-                if let Some(body) = trimmed.strip_prefix(ui::HEADER_TOOLS_PREFIX) {
-                    selected = format!("Tools: {}", body.trim());
-                    return Some(selected);
-                }
-
-                if let Some(body) = trimmed.strip_prefix(ui::HEADER_GIT_PREFIX) {
-                    let body = body.trim();
-                    if body.is_empty() {
-                        return None;
-                    }
-                    selected = body.to_string();
-                    return Some(selected);
-                }
-
-                Some(selected)
-            })
-            .collect()
-    }
-
-    fn header_meta_line(&self) -> Line<'static> {
-        let mut spans = Vec::new();
-
-        let mut first_section = true;
-        let mode_label = self.header_mode_short_label();
-        if !mode_label.trim().is_empty() {
-            spans.push(Span::styled(
-                mode_label,
-                self.header_primary_style().add_modifier(Modifier::BOLD),
-            ));
-            first_section = false;
-        }
-
-        for value in self.header_chain_values() {
-            if !first_section {
-                spans.push(Span::styled(
-                    ui::HEADER_MODE_SECONDARY_SEPARATOR.to_string(),
-                    self.header_secondary_style(),
-                ));
-            }
-            spans.push(Span::styled(value, self.header_primary_style()));
-            first_section = false;
-        }
-
-        if spans.is_empty() {
-            spans.push(Span::raw(String::new()));
-        }
-
-        Line::from(spans)
-    }
-
-    fn header_highlights_line(&self) -> Option<Line<'static>> {
-        let mut spans = Vec::new();
-        let mut first_section = true;
-
-        for highlight in &self.header_context.highlights {
-            let title = highlight.title.trim();
-            let summary = self.header_highlight_summary(highlight);
-
-            if title.is_empty() && summary.is_none() {
-                continue;
-            }
-
-            if !first_section {
-                spans.push(Span::styled(
-                    ui::HEADER_META_SEPARATOR.to_string(),
-                    self.header_secondary_style(),
-                ));
-            }
-
-            if !title.is_empty() {
-                let mut title_style = self.header_secondary_style();
-                title_style = title_style.add_modifier(Modifier::BOLD);
-                let mut title_text = title.to_string();
-                if summary.is_some() {
-                    title_text.push(':');
-                }
-                spans.push(Span::styled(title_text, title_style));
-                if summary.is_some() {
-                    spans.push(Span::styled(" ".to_string(), self.header_secondary_style()));
-                }
-            }
-
-            if let Some(body) = summary {
-                spans.push(Span::styled(body, self.header_primary_style()));
-            }
-
-            first_section = false;
-        }
-
-        if spans.is_empty() {
-            None
-        } else {
-            Some(Line::from(spans))
-        }
-    }
-
-    fn header_highlight_summary(&self, highlight: &InlineHeaderHighlight) -> Option<String> {
-        let entries: Vec<String> = highlight
-            .lines
-            .iter()
-            .map(|line| line.trim())
-            .filter(|line| !line.is_empty())
-            .map(|line| {
-                let stripped = line
-                    .strip_prefix("- ")
-                    .or_else(|| line.strip_prefix("• "))
-                    .unwrap_or(line);
-                stripped.trim().to_string()
-            })
-            .collect();
-
-        if entries.is_empty() {
-            return None;
-        }
-
-        Some(self.compact_highlight_entries(&entries))
-    }
-
-    fn compact_highlight_entries(&self, entries: &[String]) -> String {
-        let mut summary =
-            self.truncate_highlight_preview(entries.first().map(String::as_str).unwrap_or(""));
-        if entries.len() > 1 {
-            let remaining = entries.len() - 1;
-            if !summary.is_empty() {
-                let _ = write!(summary, " (+{} more)", remaining);
-            } else {
-                summary = format!("(+{} more)", remaining);
-            }
-        }
-        summary
-    }
-
-    fn truncate_highlight_preview(&self, text: &str) -> String {
-        let max = ui::HEADER_HIGHLIGHT_PREVIEW_MAX_CHARS;
-        if max == 0 {
-            return String::new();
-        }
-
-        let grapheme_count = text.graphemes(true).count();
-        if grapheme_count <= max {
-            return text.to_string();
-        }
-
-        let mut truncated = String::new();
-        for grapheme in text.graphemes(true).take(max.saturating_sub(1)) {
-            truncated.push_str(grapheme);
-        }
-        truncated.push_str(ui::INLINE_PREVIEW_ELLIPSIS);
-        truncated
-    }
-
-    /// Determine if suggestions should be shown in the header
-    fn should_show_suggestions(&self) -> bool {
-        // Show suggestions when input is empty or starts with /
-        self.input.is_empty() || self.input.starts_with('/')
-    }
-
-    /// Generate header line with slash command and keyboard shortcut suggestions
-    fn header_suggestions_line(&self) -> Option<Line<'static>> {
-        let mut spans = Vec::new();
-
-        spans.push(Span::styled(
-            "/help",
-            self.header_primary_style().add_modifier(Modifier::BOLD),
-        ));
-        spans.push(Span::styled(
-            " · ",
-            self.header_secondary_style().add_modifier(Modifier::DIM),
-        ));
-        spans.push(Span::styled(
-            format!("/{}", PROMPT_COMMAND_NAME),
-            self.header_primary_style().add_modifier(Modifier::BOLD),
-        ));
-        spans.push(Span::styled(
-            " · ",
-            self.header_secondary_style().add_modifier(Modifier::DIM),
-        ));
-        spans.push(Span::styled(
-            "/model",
-            self.header_primary_style().add_modifier(Modifier::BOLD),
-        ));
-        spans.push(Span::styled(
-            "  |  ",
-            self.header_secondary_style().add_modifier(Modifier::DIM),
-        ));
-        spans.push(Span::styled(
-            "↑↓",
-            self.header_primary_style().add_modifier(Modifier::BOLD),
-        ));
-        spans.push(Span::styled(" Nav · ", self.header_secondary_style()));
-        spans.push(Span::styled(
-            "Tab",
-            self.header_primary_style().add_modifier(Modifier::BOLD),
-        ));
-        spans.push(Span::styled(" Complete", self.header_secondary_style()));
-
-        Some(Line::from(spans))
-    }
-
-    fn section_title_style(&self) -> Style {
-        let mut style = self.default_style().add_modifier(Modifier::BOLD);
-        if let Some(primary) = self.theme.primary.or(self.theme.foreground) {
-            style = style.fg(ratatui_color_from_ansi(primary));
-        }
-        style
-    }
-
-    fn header_primary_style(&self) -> Style {
-        let mut style = self.default_style();
-        if let Some(primary) = self.theme.primary.or(self.theme.foreground) {
-            style = style.fg(ratatui_color_from_ansi(primary));
-        }
-        style
-    }
-
-    fn header_secondary_style(&self) -> Style {
-        let mut style = self.default_style();
-        if let Some(secondary) = self.theme.secondary.or(self.theme.foreground) {
-            style = style.fg(ratatui_color_from_ansi(secondary));
-        }
-        style
-    }
-
-    fn suggestion_block_title(&self) -> Line<'static> {
-        Line::from(vec![Span::styled(
-            ui::SUGGESTION_BLOCK_TITLE.to_string(),
-            self.section_title_style(),
-        )])
-    }
-
-    fn navigation_block_title(&self) -> Line<'static> {
-        if self.should_show_plan() {
-            return self.plan_block_title();
-        }
-
-        let mut spans = Vec::new();
-        spans.push(Span::styled(
-            ui::NAVIGATION_BLOCK_TITLE.to_string(),
-            self.section_title_style(),
-        ));
-        spans.push(Span::styled(
-            format!(" · {}", ui::NAVIGATION_BLOCK_SHORTCUT_NOTE),
-            self.navigation_preview_style(),
-        ));
-
-        Line::from(spans)
-    }
-
-    fn plan_block_title(&self) -> Line<'static> {
-        let mut spans = Vec::new();
-        spans.push(Span::styled(
-            ui::PLAN_BLOCK_TITLE.to_string(),
-            self.section_title_style(),
-        ));
-
-        let status = self.plan_status_label();
-        spans.push(Span::styled(
-            format!(" · {}", status),
-            self.navigation_preview_style(),
-        ));
-
-        if self.plan.summary.total_steps > 0 {
-            spans.push(Span::styled(
-                format!(
-                    " · {}/{}",
-                    self.plan.summary.completed_steps, self.plan.summary.total_steps
-                ),
-                self.navigation_preview_style(),
-            ));
-        }
-
-        Line::from(spans)
-    }
-
-    fn plan_status_label(&self) -> &'static str {
-        match self.plan.summary.status {
-            PlanCompletionState::Done => ui::PLAN_STATUS_DONE,
-            PlanCompletionState::InProgress => ui::PLAN_STATUS_IN_PROGRESS,
-            PlanCompletionState::Empty => ui::PLAN_STATUS_EMPTY,
-        }
-    }
-
-    fn navigation_items(&self) -> Vec<ListItem<'static>> {
-        if self.should_show_plan() {
-            return self.plan_navigation_items();
-        }
-        self.timeline_navigation_items()
-    }
-
-    fn timeline_navigation_items(&self) -> Vec<ListItem<'static>> {
-        if self.lines.is_empty() {
-            return vec![ListItem::new(Line::from(vec![Span::styled(
-                ui::NAVIGATION_EMPTY_LABEL.to_string(),
-                self.navigation_placeholder_style(),
-            )]))];
-        }
-
-        self.lines
-            .iter()
-            .enumerate()
-            .map(|(index, line)| ListItem::new(Line::from(self.navigation_spans(index, line))))
-            .collect()
-    }
-
-    fn plan_navigation_items(&self) -> Vec<ListItem<'static>> {
-        self.plan
-            .steps
-            .iter()
-            .enumerate()
-            .map(|(index, step)| ListItem::new(Line::from(self.plan_step_spans(index, step))))
-            .collect()
-    }
-
-    fn plan_step_spans(&self, index: usize, step: &PlanStep) -> Vec<Span<'static>> {
-        let mut spans = Vec::new();
-        let sequence = format!("{}{:02}", ui::NAVIGATION_INDEX_PREFIX, index + 1);
-        spans.push(Span::styled(sequence, self.navigation_index_style()));
-        spans.push(Span::raw(" "));
-        spans.push(Span::styled(
-            step.status.checkbox().to_string(),
-            self.plan_checkbox_style(step.status.clone()),
-        ));
-        spans.push(Span::raw(" "));
-        spans.push(Span::styled(
-            step.step.clone(),
-            self.plan_step_style(step.status.clone()),
-        ));
-        if matches!(step.status, StepStatus::InProgress) {
-            spans.push(Span::raw(" "));
-            spans.push(Span::styled(
-                format!("({})", ui::PLAN_IN_PROGRESS_NOTE),
-                self.plan_status_note_style(),
-            ));
-        }
-        spans
-    }
-
-    fn navigation_spans(&self, index: usize, line: &MessageLine) -> Vec<Span<'static>> {
-        let mut spans = Vec::new();
-        let sequence = format!("{}{:02}", ui::NAVIGATION_INDEX_PREFIX, index + 1);
-        spans.push(Span::styled(sequence, self.navigation_index_style()));
-        spans.push(Span::raw(" "));
-        spans.push(Span::styled(
-            self.navigation_label(line.kind).to_string(),
-            self.navigation_label_style(line.kind),
-        ));
-        let preview = self.navigation_preview_text(line);
-        if !preview.is_empty() {
-            spans.push(Span::raw(" "));
-            spans.push(Span::styled(preview, self.navigation_preview_style()));
-        }
-        spans
-    }
-
-    fn navigation_label(&self, kind: InlineMessageKind) -> &'static str {
-        match kind {
-            InlineMessageKind::Agent => ui::NAVIGATION_LABEL_AGENT,
-            InlineMessageKind::Error => ui::NAVIGATION_LABEL_ERROR,
-            InlineMessageKind::Info => ui::NAVIGATION_LABEL_INFO,
-            InlineMessageKind::Policy => ui::NAVIGATION_LABEL_POLICY,
-            InlineMessageKind::Tool => ui::NAVIGATION_LABEL_TOOL,
-            InlineMessageKind::User => ui::NAVIGATION_LABEL_USER,
-            InlineMessageKind::Pty => ui::NAVIGATION_LABEL_PTY,
-        }
-    }
-
-    fn navigation_preview_text(&self, line: &MessageLine) -> String {
-        let mut preview = String::new();
-        let mut char_count = 0usize;
-        let mut truncated = false;
-        for segment in &line.segments {
-            let sanitized = segment.text.replace('\n', " ").replace('\r', " ");
-            let trimmed = sanitized.trim();
-            if trimmed.is_empty() {
-                continue;
-            }
-            if char_count > 0 {
-                if char_count + 1 > ui::INLINE_PREVIEW_MAX_CHARS {
-                    truncated = true;
-                    break;
-                }
-                preview.push(' ');
-                char_count += 1;
-            }
-            for character in trimmed.chars() {
-                if char_count == ui::INLINE_PREVIEW_MAX_CHARS {
-                    truncated = true;
-                    break;
-                }
-                preview.push(character);
-                char_count += 1;
-            }
-            if truncated {
-                break;
-            }
-        }
-
-        if truncated {
-            preview.push_str(ui::INLINE_PREVIEW_ELLIPSIS);
-        }
-
-        preview
-    }
-
-    fn navigation_index_style(&self) -> Style {
-        self.header_secondary_style().add_modifier(Modifier::DIM)
-    }
-
-    fn navigation_label_style(&self, kind: InlineMessageKind) -> Style {
-        let mut style = InlineTextStyle::default();
-        style.color = self.text_fallback(kind).or(self.theme.foreground);
-        style.bold = matches!(kind, InlineMessageKind::Agent | InlineMessageKind::User);
-        ratatui_style_from_inline(&style, self.theme.foreground)
-    }
-
-    fn navigation_preview_style(&self) -> Style {
-        self.default_style().add_modifier(Modifier::DIM)
-    }
-
-    fn navigation_placeholder_style(&self) -> Style {
-        self.default_style().add_modifier(Modifier::DIM)
-    }
-
-    fn plan_checkbox_style(&self, status: StepStatus) -> Style {
-        match status {
-            StepStatus::Completed => self.navigation_preview_style(),
-            StepStatus::InProgress => self.accent_style().add_modifier(Modifier::BOLD),
-            StepStatus::Pending => self.default_style(),
-        }
-    }
-
-    fn plan_step_style(&self, status: StepStatus) -> Style {
-        match status {
-            StepStatus::Completed => self.navigation_preview_style(),
-            StepStatus::InProgress => self.accent_style().add_modifier(Modifier::BOLD),
-            StepStatus::Pending => self.default_style(),
-        }
-    }
-
-    fn plan_status_note_style(&self) -> Style {
-        self.navigation_preview_style()
-    }
-
-    fn navigation_highlight_style(&self) -> Style {
-        let mut style = Style::default().add_modifier(Modifier::REVERSED | Modifier::BOLD);
-        if let Some(primary) = self.theme.primary.or(self.theme.secondary) {
-            style = style.fg(ratatui_color_from_ansi(primary));
-        }
-        style
-    }
-
-    fn plan_selected_index(&self) -> Option<usize> {
-        if self.plan.steps.is_empty() {
-            return None;
-        }
-
-        if let Some(index) = self
-            .plan
-            .steps
-            .iter()
-            .position(|step| matches!(step.status, StepStatus::InProgress))
-        {
-            return Some(index);
-        }
-
-        if let Some(index) = self
-            .plan
-            .steps
-            .iter()
-            .position(|step| matches!(step.status, StepStatus::Pending))
-        {
-            return Some(index);
-        }
-
-        Some(self.plan.steps.len().saturating_sub(1))
-    }
-
-    fn should_show_plan(&self) -> bool {
-        self.plan.summary.status != PlanCompletionState::Empty && !self.plan.steps.is_empty()
+    fn set_plan(&mut self, plan: TaskPlan) {
+        self.plan = plan;
+        self.mark_dirty();
     }
 
     fn modal_list_highlight_style(&self) -> Style {
@@ -1374,94 +608,18 @@ impl Session {
         frame.render_widget(paragraph, scroll_area);
     }
 
-    fn set_plan(&mut self, plan: TaskPlan) {
-        self.plan = plan;
-        self.mark_dirty();
+    fn header_reserved_rows(&self) -> u16 {
+        self.header_rows.max(ui::INLINE_HEADER_HEIGHT)
     }
 
-    fn render_slash_palette(&mut self, frame: &mut Frame<'_>, viewport: Rect) {
-        if viewport.height == 0 || viewport.width == 0 || self.modal.is_some() {
-            self.slash_palette.clear_visible_rows();
-            return;
-        }
-        let suggestions = self.slash_palette.suggestions();
-        if suggestions.is_empty() {
-            self.slash_palette.clear_visible_rows();
-            return;
-        }
-
-        let mut width_hint = measure_text_width(ui::SLASH_PALETTE_HINT_PRIMARY);
-        width_hint = width_hint.max(measure_text_width(ui::SLASH_PALETTE_HINT_SECONDARY));
-        for suggestion in suggestions.iter().take(ui::SLASH_SUGGESTION_LIMIT) {
-            let label = match suggestion {
-                slash_palette::SlashPaletteSuggestion::Static(cmd) => {
-                    if !cmd.description.is_empty() {
-                        format!("/{} {}", cmd.name, cmd.description)
-                    } else {
-                        format!("/{}", cmd.name)
-                    }
-                }
-                slash_palette::SlashPaletteSuggestion::Custom(prompt) => {
-                    // For custom prompts, format as /prompt:name (legacy alias /prompts:name)
-                    let prompt_cmd = format!("{}:{}", PROMPT_COMMAND_NAME, prompt.name);
-                    let description = prompt.description.as_deref().unwrap_or("");
-                    if !description.is_empty() {
-                        format!("/{} {}", prompt_cmd, description)
-                    } else {
-                        format!("/{}", prompt_cmd)
-                    }
-                }
-            };
-            width_hint = width_hint.max(measure_text_width(&label));
-        }
-
-        let instructions = self.slash_palette_instructions();
-        let area = compute_modal_area(viewport, width_hint, instructions.len(), 0, 0, true);
-
-        frame.render_widget(Clear, area);
-        let block = Block::default()
-            .title(self.suggestion_block_title())
-            .borders(Borders::ALL)
-            .border_type(BorderType::Rounded)
-            .style(self.default_style())
-            .border_style(self.border_style());
-        let inner = block.inner(area);
-        frame.render_widget(block, area);
-        if inner.height == 0 || inner.width == 0 {
-            self.slash_palette.clear_visible_rows();
-            return;
-        }
-
-        let layout = ModalListLayout::new(inner, instructions.len());
-        if let Some(text_area) = layout.text_area {
-            let paragraph = Paragraph::new(instructions).wrap(Wrap { trim: true });
-            frame.render_widget(paragraph, text_area);
-        }
-
-        self.slash_palette
-            .set_visible_rows(layout.list_area.height as usize);
-
-        // Get all list items (scrollable via ListState)
-        let list_items = self.slash_list_items();
-
-        let list = List::new(list_items)
-            .style(self.default_style())
-            .highlight_style(self.slash_highlight_style());
-
-        frame.render_stateful_widget(list, layout.list_area, self.slash_palette.list_state_mut());
+    fn input_reserved_rows(&self) -> u16 {
+        self.header_reserved_rows() + self.input_height
     }
 
-    fn slash_palette_instructions(&self) -> Vec<Line<'static>> {
-        vec![
-            Line::from(Span::styled(
-                ui::SLASH_PALETTE_HINT_PRIMARY.to_string(),
-                self.default_style(),
-            )),
-            Line::from(Span::styled(
-                ui::SLASH_PALETTE_HINT_SECONDARY.to_string(),
-                self.default_style().add_modifier(Modifier::DIM),
-            )),
-        ]
+    fn recalculate_transcript_rows(&mut self) {
+        let reserved = self.input_reserved_rows().saturating_add(2); // account for transcript block borders
+        let available = self.view_rows.saturating_sub(reserved).max(1);
+        self.apply_transcript_rows(available);
     }
 
     fn render_file_palette(&mut self, frame: &mut Frame<'_>, viewport: Rect) {
@@ -1814,339 +972,6 @@ impl Session {
         self.input_status_right
             .as_ref()
             .is_some_and(|value| !value.trim().is_empty())
-    }
-
-    fn slash_list_items(&self) -> Vec<ListItem<'static>> {
-        let command_style = self.slash_name_style();
-        let match_style = self.slash_match_style();
-        let description_style = self.slash_description_style();
-
-        self.slash_palette
-            .items()
-            .into_iter()
-            .map(|item| {
-                let mut spans: Vec<Span<'static>> = Vec::new();
-                spans.push(Span::styled("/".to_string(), command_style));
-                spans.extend(item.name_segments.into_iter().map(|segment| {
-                    let style = if segment.highlighted {
-                        match_style
-                    } else {
-                        command_style
-                    };
-                    Span::styled(segment.content, style)
-                }));
-                if !item.description.is_empty() {
-                    spans.push(Span::raw(" "));
-                    spans.push(Span::styled(
-                        item.description.to_string(),
-                        description_style,
-                    ));
-                }
-                ListItem::new(Line::from(spans))
-            })
-            .collect()
-    }
-
-    fn slash_match_style(&self) -> Style {
-        self.slash_name_style().add_modifier(Modifier::UNDERLINED)
-    }
-
-    fn slash_highlight_style(&self) -> Style {
-        let highlight = self
-            .theme
-            .primary
-            .or(self.theme.secondary)
-            .or(self.theme.foreground);
-        let mut style = Style::default().add_modifier(Modifier::BOLD | Modifier::REVERSED);
-        if let Some(color) = highlight {
-            style = style.fg(ratatui_color_from_ansi(color));
-        }
-        style
-    }
-
-    fn slash_name_style(&self) -> Style {
-        let color = self.theme.primary.or(self.theme.foreground);
-        let mut style = Style::default().add_modifier(Modifier::BOLD);
-        if let Some(color) = color {
-            style = style.fg(ratatui_color_from_ansi(color));
-        }
-        style
-    }
-
-    fn slash_description_style(&self) -> Style {
-        let color = self.theme.secondary.or(self.theme.foreground);
-        let mut style = Style::default().add_modifier(Modifier::DIM);
-        if let Some(color) = color {
-            style = style.fg(ratatui_color_from_ansi(color));
-        }
-        style
-    }
-
-    fn header_reserved_rows(&self) -> u16 {
-        self.header_rows.max(ui::INLINE_HEADER_HEIGHT)
-    }
-
-    fn input_reserved_rows(&self) -> u16 {
-        self.header_reserved_rows() + self.input_height
-    }
-
-    fn recalculate_transcript_rows(&mut self) {
-        let reserved = self.input_reserved_rows().saturating_add(2); // account for transcript block borders
-        let available = self.view_rows.saturating_sub(reserved).max(1);
-        self.apply_transcript_rows(available);
-    }
-
-    fn handle_slash_palette_change(&mut self) {
-        self.recalculate_transcript_rows();
-        self.enforce_scroll_bounds();
-        self.mark_dirty();
-    }
-
-    fn clear_slash_suggestions(&mut self) {
-        if self.slash_palette.clear() {
-            self.handle_slash_palette_change();
-        }
-    }
-
-    fn update_slash_suggestions(&mut self) {
-        if !self.input_enabled {
-            self.clear_slash_suggestions();
-            return;
-        }
-
-        let Some(prefix) = command_prefix(&self.input, self.cursor) else {
-            self.clear_slash_suggestions();
-            return;
-        };
-
-        // Update slash palette with custom prompts if available
-        if let Some(ref custom_prompts) = self.custom_prompts {
-            self.slash_palette
-                .set_custom_prompts(custom_prompts.clone());
-        }
-
-        match self
-            .slash_palette
-            .update(Some(&prefix), ui::SLASH_SUGGESTION_LIMIT)
-        {
-            SlashPaletteUpdate::NoChange => {}
-            SlashPaletteUpdate::Cleared | SlashPaletteUpdate::Changed { .. } => {
-                self.handle_slash_palette_change();
-            }
-        }
-    }
-
-    fn slash_navigation_available(&self) -> bool {
-        self.input_enabled && !self.slash_palette.is_empty()
-    }
-
-    fn move_slash_selection_up(&mut self) -> bool {
-        let changed = self.slash_palette.move_up();
-        self.handle_slash_selection_change(changed)
-    }
-
-    fn move_slash_selection_down(&mut self) -> bool {
-        let changed = self.slash_palette.move_down();
-        self.handle_slash_selection_change(changed)
-    }
-
-    fn select_first_slash_suggestion(&mut self) -> bool {
-        let changed = self.slash_palette.select_first();
-        self.handle_slash_selection_change(changed)
-    }
-
-    fn select_last_slash_suggestion(&mut self) -> bool {
-        let changed = self.slash_palette.select_last();
-        self.handle_slash_selection_change(changed)
-    }
-
-    fn page_up_slash_suggestion(&mut self) -> bool {
-        let changed = self.slash_palette.page_up();
-        self.handle_slash_selection_change(changed)
-    }
-
-    fn page_down_slash_suggestion(&mut self) -> bool {
-        let changed = self.slash_palette.page_down();
-        self.handle_slash_selection_change(changed)
-    }
-
-    fn handle_slash_selection_change(&mut self, changed: bool) -> bool {
-        if changed {
-            self.preview_selected_slash_suggestion();
-            self.recalculate_transcript_rows();
-            self.enforce_scroll_bounds();
-            self.mark_dirty();
-            true
-        } else {
-            false
-        }
-    }
-
-    fn preview_selected_slash_suggestion(&mut self) {
-        let Some(command) = self.slash_palette.selected_command() else {
-            return;
-        };
-        let Some(range) = command_range(&self.input, self.cursor) else {
-            return;
-        };
-
-        let current_input = self.input.clone();
-        let prefix = &current_input[..range.start];
-        let suffix = &current_input[range.end..];
-
-        let mut new_input = String::new();
-        new_input.push_str(prefix);
-        new_input.push('/');
-        new_input.push_str(command.name);
-        let cursor_position = new_input.len();
-
-        if !suffix.is_empty() {
-            if !suffix.chars().next().map_or(false, char::is_whitespace) {
-                new_input.push(' ');
-            }
-            new_input.push_str(suffix);
-        }
-
-        self.input = new_input;
-        self.cursor = cursor_position.min(self.input.len());
-        self.mark_dirty();
-    }
-
-    fn apply_selected_slash_suggestion(&mut self) -> bool {
-        // Check if there's a selected custom prompt first
-        if let Some(custom_prompt) = self.slash_palette.selected_custom_prompt() {
-            let Some(range) = command_range(&self.input, self.cursor) else {
-                return false;
-            };
-
-            // Replace the input with the selected custom prompt in /prompt:name format
-            let mut new_input = String::from(PROMPT_COMMAND_PREFIX);
-            new_input.push_str(&custom_prompt.name);
-
-            let suffix = &self.input[range.end..];
-            if !suffix.is_empty() {
-                if !suffix.chars().next().map_or(false, char::is_whitespace) {
-                    new_input.push(' ');
-                }
-                new_input.push_str(suffix);
-            } else {
-                new_input.push(' ');
-            }
-
-            let cursor_position = new_input.len();
-
-            self.input = new_input;
-            self.cursor = cursor_position;
-            self.update_slash_suggestions();
-            self.mark_dirty();
-            return true;
-        }
-
-        // Fall back to regular command if no custom prompt is selected
-        let Some(command) = self.slash_palette.selected_command() else {
-            return false;
-        };
-
-        // Store command name before borrowing self mutably
-        let command_name = command.name.to_string();
-
-        let Some(range) = command_range(&self.input, self.cursor) else {
-            return false;
-        };
-
-        let suffix = self.input[range.end..].to_string();
-        let mut new_input = format!("/{}", command_name);
-
-        let cursor_position = if suffix.is_empty() {
-            new_input.push(' ');
-            new_input.len()
-        } else {
-            if !suffix.chars().next().map_or(false, char::is_whitespace) {
-                new_input.push(' ');
-            }
-            let position = new_input.len();
-            new_input.push_str(&suffix);
-            position
-        };
-
-        self.input = new_input;
-        self.cursor = cursor_position;
-
-        // Special handling: if /files command was selected, defer file browser opening
-        if command_name == "files" {
-            // Clear slash modal first
-            self.clear_slash_suggestions();
-            self.mark_dirty();
-
-            // Set flag to open file browser on next render cycle (after modal fully dismisses)
-            self.deferred_file_browser_trigger = true;
-        } else if command_name == PROMPT_COMMAND_NAME || command_name == LEGACY_PROMPT_COMMAND_NAME
-        {
-            // Clear slash modal first
-            self.clear_slash_suggestions();
-            self.mark_dirty();
-
-            // Set flag to open prompt browser on next render cycle (after modal fully dismisses)
-            self.deferred_prompt_browser_trigger = true;
-        } else {
-            // For other commands, update slash suggestions normally
-            self.update_slash_suggestions();
-            self.mark_dirty();
-        }
-
-        true
-    }
-
-    fn try_handle_slash_navigation(
-        &mut self,
-        key: &KeyEvent,
-        has_control: bool,
-        has_alt: bool,
-        has_command: bool,
-    ) -> bool {
-        if !self.slash_navigation_available() || has_control || has_alt {
-            return false;
-        }
-
-        let handled = match key.code {
-            KeyCode::Up => {
-                if has_command {
-                    self.select_first_slash_suggestion()
-                } else {
-                    self.move_slash_selection_up()
-                }
-            }
-            KeyCode::Down => {
-                if has_command {
-                    self.select_last_slash_suggestion()
-                } else {
-                    self.move_slash_selection_down()
-                }
-            }
-            KeyCode::PageUp => self.page_up_slash_suggestion(),
-            KeyCode::PageDown => self.page_down_slash_suggestion(),
-            KeyCode::Home => self.select_first_slash_suggestion(),
-            KeyCode::End => self.select_last_slash_suggestion(),
-            KeyCode::Tab => self.apply_selected_slash_suggestion(),
-            KeyCode::BackTab => self.move_slash_selection_up(),
-            _ => return false,
-        };
-
-        if handled {
-            true
-        } else {
-            matches!(
-                key.code,
-                KeyCode::Up
-                    | KeyCode::Down
-                    | KeyCode::PageUp
-                    | KeyCode::PageDown
-                    | KeyCode::Home
-                    | KeyCode::End
-                    | KeyCode::Tab
-                    | KeyCode::BackTab
-            )
-        }
     }
 
     fn render_message_spans(&self, index: usize) -> Vec<Span<'static>> {
